feat: refactored key chain delete
